// src/app/models/popular-activity.ts
export class PopularActivity {
    added_date: string;
    category: string;
    city: string;
    country: string;
    cover: string;
    currency: string;
    end_date: string;
    end_time: string;
    id: string;
    is_active: boolean;
    is_free: boolean;
    is_hosting: boolean;
    is_wishlist: boolean;
    picture: string;
    source_link: string;
    start_date: string;
    start_time: string;
    title: string;
    body: string;
    source: string;
    saved: boolean;
    summary: string;
}